requirements.txt
groq
python-dotenv
flask
flask-cors
edge-tts
